DF-796 - extract request-mapper, prefill with template value, adjust translations
