+++ alv-portal-ui/src/app/job-advertisement/job-ad-search/job-detail/complaint-modal/complaint-modal.component.html
@@ -1,18 +1,18 @@
-<alv-modal modalTitle="complain-modal.header"
-           primaryLabel="complain-modal.send"
+<alv-modal modalTitle="complaint-modal.header"
+           primaryLabel="complaint-modal.send"
            (primaryAction)="onSubmit(form)"
-           secondaryLabel="complain-modal.dismiss"
+           secondaryLabel="complaint-modal.dismiss"
            (secondaryAction)="activeModal.dismiss()"
            [formGroup]="form">
-  <p [innerHTML]="'complain-modal.description' | translate"></p>
+  <p [innerHTML]="'complaint-modal.description' | translate"></p>
   <fieldset>
-    <p><strong>{{'complain-modal.contact.description' | translate}}</strong></p>
+    <p><strong>{{'complaint-modal.contact.description' | translate}}</strong></p>
     <alv-select label="home.tools.job-publication.contact.salutation"
                 alvFormControlName="salutation"
                 [options$]="salutationOptions$">
     </alv-select>
 
-    <alv-input-field label="complain-modal.name-lastName.label"
+    <alv-input-field label="complaint-modal.name-lastName.label"
                      alvFormControlName="name"
                      [maxLength]="MAX_LENGTH_255">
     </alv-input-field>
@@ -30,8 +30,8 @@
     </alv-input-field>
   </fieldset>
   <fieldset>
-    <p><strong>{{'complain-modal.message.label' | translate}}</strong></p>
-    <alv-input-field label="complain-modal.message.placeholder"
+    <p><strong>{{'complaint-modal.message.label' | translate}}</strong></p>
+    <alv-input-field label="complaint-modal.message.placeholder"
                      alvFormControlName="complaintMessage"
                      [maxLength]="MAX_LENGTH_1000"
                      [showCharacterCounter]="true"
