+++ alv-portal-ui/src/app/job-advertisement/job-ad-search/job-detail/complaint-modal/complaint-modal.component.ts
@@ -1,11 +1,15 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { Observable, of } from 'rxjs';
+import { combineLatest, Observable, of } from 'rxjs';
 import { Salutation } from '../../../../shared/backend-services/shared.types';
 import { ComplaintRepository } from '../../../../shared/backend-services/complaint/complaint.repository';
 import { ComplaintDto } from '../../../../shared/backend-services/complaint/complaint.types';
-import { switchMap } from 'rxjs/operators';
+import { switchMap, takeUntil } from 'rxjs/operators';
+import { AuthenticationService } from '../../../../core/auth/authentication.service';
+import { CompanyContactTemplateModel } from '../../../../core/auth/company-contact-template-model';
+import { AbstractSubscriber } from '../../../../core/abstract-subscriber';
+import { mapFormToDto, mapToFormValue } from './complaint-request-mapper';
 
 export interface ComplaintFormValue {
   salutation: Salutation;
@@ -19,7 +23,7 @@ export interface ComplaintFormValue {
   selector: 'alv-complaint-modal',
   templateUrl: './complaint-modal.component.html',
 })
-export class ComplaintModalComponent implements OnInit {
+export class ComplaintModalComponent extends AbstractSubscriber implements OnInit {
 
   public form: FormGroup;
 
@@ -46,12 +50,15 @@ export class ComplaintModalComponent implements OnInit {
   );
 
   constructor(public activeModal: NgbActiveModal,
+              private authenticationService: AuthenticationService,
               private fb: FormBuilder,
               private complaintRepository: ComplaintRepository) {
+    super();
   }
 
   ngOnInit() {
     const formValue = mapToFormValue(this.complaint);
+
     this.form = this.fb.group({
       salutation: [formValue.salutation, Validators.required],
       name: [formValue.name, [Validators.required, Validators.maxLength(this.MAX_LENGTH_255)]],
@@ -59,6 +66,14 @@ export class ComplaintModalComponent implements OnInit {
       email: [formValue.email, Validators.required],
       complaintMessage: [formValue.complaintMessage, [Validators.required, Validators.maxLength(this.MAX_LENGTH_1000)]]
     });
+
+    combineLatest(this.authenticationService.getCurrentCompany()).pipe(
+      takeUntil(this.ngUnsubscribe))
+      .subscribe(([templateInfo]) => {
+        if (!!templateInfo) {
+          this.patchTemplateValues(templateInfo);
+        }
+      });
   }
 
   onSubmit(form: FormGroup) {
@@ -67,27 +82,14 @@ export class ComplaintModalComponent implements OnInit {
       switchMap((id) => this.complaintRepository.sendComplaint(mapFormToDto(id, formValue)))
     ).subscribe(() => this.activeModal.close());
   }
-}
 
-function mapToFormValue(complaint: ComplaintDto): ComplaintFormValue {
-  return {
-    complaintMessage: complaint.complaintMessage,
-    salutation: complaint.contactInformation.salutation,
-    name: complaint.contactInformation.name,
-    phone: complaint.contactInformation.phone,
-    email: complaint.contactInformation.email
-  };
-}
+  private patchTemplateValues(templateInfo: CompanyContactTemplateModel): void {
+    this.form.patchValue({
+      salutation: templateInfo.salutation,
+      name: templateInfo.companyName,
+      phone: templateInfo.phone,
+      email: templateInfo.email
+    });
+  }
 
-export function mapFormToDto(id: string, formValue: ComplaintFormValue): ComplaintDto {
-  return {
-    jobAdvertisementId: id,
-    contactInformation: {
-      salutation: formValue.salutation,
-      name: formValue.name,
-      phone: formValue.phone,
-      email: formValue.email,
-    },
-    complaintMessage: formValue.complaintMessage
-  };
 }
