+++ alv-portal-ui/translations.csv
@@ -527,15 +527,14 @@ job-detail.startsImmediately.false,Nach Vereinbarung,By agreement,À convenir,Pr
 job-detail.startsImmediately.label,Stellenantritt,Start date,Entrée en fonction,Inizio dell'attività,
 job-detail.startsImmediately.true,Ab Sofort,Immediately,Tout de suite,Da subito,
 job-detail.unvalidated,Dieses Inserat wurde direkt von einem unserer Nutzer elektronisch auf die Internetplattform der öffentlichen Arbeitsvermittlung publiziert. Die Arbeitslosenversicherung und öffentliche Arbeitsvermittlung haben keinen Einfluss auf dessen Inhalt und Qualität.,This advert was published electronically by one of our users on the internet platform of the public employment service. Neither the unemployment insurance office nor the public employment service have any influence over its content or quality.,Cette offre d’emploi a été publiée en ligne sur la plateforme Internet du service public de l’emploi directement par l’un de nos utilisateurs. L’assurance-chômage et le service public de l’emploi ne sont pas responsables de son contenu et de sa qualité.,Questo annuncio è stato pubblicato in formato elettronico da uno dei nostri utenti direttamente sulla piattaforma online del servizio pubblico di collocamento. L’assicurazione contro la disoccupazione e il servizio pubblico di collocamento non rispondono né dei contenuti né della qualità.,
-complain-modal.header,Stelle beanstanden,,,,
-complain-modal.description,"Bitte melden Sie uns unseriöse Stellenanzeigen.<br/>Diese erkennen Sie z.B. an folgenden Eigenschaften:<ul><li>Unvollständige Kontaktdaten</li><li>Vorab Investitionen</li><li>Sittenwidrige Inhalte</li><li>Diskriminierende Inhalte</li><li>Sexistische Inhalte</li></ul>",,,,
-complain-modal.contact.description,Ihre Kontaktdaten,,,,
-
-complain-modal.name-lastName.label,Vorname / Name,First name / Surname,Prénom / Nom,Nome / Cognome,
-complain-modal.send,Absenden,Submit,Soumettre,Inviare,
-complain-modal.dismiss,Abbrechen,Cancel,Annuler,Annullare,
-complain-modal.message.label,"Warum möchten Sie diese Stelle beanstanden?",,,,
-complain-modal.message.placeholder,Ihre persönliche Nachricht an uns,,,,
+complaint-modal.header,Stelle beanstanden,,,,
+complaint-modal.description,"Bitte melden Sie uns unseriöse Stellenanzeigen.<br/>Diese erkennen Sie z.B. an folgenden Eigenschaften:<ul><li>Unvollständige Kontaktdaten</li><li>Vorab Investitionen</li><li>Sittenwidrige Inhalte</li><li>Diskriminierende Inhalte</li><li>Sexistische Inhalte</li></ul>",,,,
+complaint-modal.contact.description,Ihre Kontaktdaten,,,,
+complaint-modal.name-lastName.label,Vorname / Name,First name / Surname,Prénom / Nom,Nome / Cognome,
+complaint-modal.send,Absenden,Submit,Soumettre,Inviare,
+complaint-modal.dismiss,Abbrechen,Cancel,Annuler,Annullare,
+complaint-modal.message.label,"Warum möchten Sie diese Stelle beanstanden?",,,,
+complaint-modal.message.placeholder,Ihre persönliche Nachricht an uns,,,,
 job-publication-cancel.job.change-or-repost,Sie möchten die Ausschreibung ändern oder neu erfassen,You want to change or repost the position,Vous souhaitez modifier ou republier l'annonce?,Desiderate modificare o registrare nuovamente la pubblicazione ,
 job-publication-cancel.job.confirm,Abmeldung bestätigen,Confirm cancellation,Confirmer l'annulation,Confermare la cancellazione,
 job-publication-cancel.job.dismiss,Abbrechen,Cancel,Annuler,Annullare,
